python_solution/requirements.txt
numpy
pillow
opencv-python
matplotlib
scipy
plotly
